@@ -5,16 +5,16 @@
   "dependencies": {
     "@formspree/react": "^2.2.4",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.2.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "@vercel/analytics": "^1.4.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-animate-on-scroll": "^2.1.5",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-google-recaptcha": "^2.1.0",
     "react-router-dom": "^6.2.1",
     "react-router-hash-link": "^2.4.3",
